+++ crates/runtime/src/flight/flightsql.rs
@@ -1,0 +1,6 @@
+pub(crate) mod get_catalogs;
+pub(crate) mod get_schemas;
+pub(crate) mod get_sql_info;
+pub(crate) mod get_tables;
+pub(crate) mod prepared_statement_query;
+pub(crate) mod statement_query;
